Updates Snehal xgboost model with higher r^2 value
This model essentially uses the XGBoost model that Snehal created and we modified some metrics to drop to increase the r^2 to 0.593 from 0.586. In addition, this creates the performance index of each player (0-100) and provides the predicted salary, compares to their actual salary, the salary difference, and if the player is underpaid/overpaid/correctly paid.